requirements.txt
asgiref==3.8.1
Django==5.1.7
djangorestframework==3.15.2
et_xmlfile==2.0.0
openpyxl==3.1.5
RapidFuzz==3.12.2
sqlparse==0.5.3
thefuzz==0.22.1
tzdata==2025.1
